Add -UseBasicParsing to invoke command
Unless `-UseBasicParsing` is specificed, `Invoke-WebRequest` uses
Internet Explorer under the hood to build a DOM of the response. If
IE hasn't been run yet, its first-run setup isn't done, and
`Invoke-WebRequest` fails.

To fix this, add `-UseBasicParsing` since the command is just
donwloading a file and not doing DOM inspection.